requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
aiosmtplib
python-multipart
python-dotenv
httpx
